requirements.txt
absl-py==2.1.0
ale-py==0.8.1
astunparse==1.6.3
attrs==23.2.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
blackjax==0.3.0
box2d-py==2.3.5
certifi==2024.6.2
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
coax==0.1.13
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
dataclasses==0.6
dill==0.3.8
dm-env==1.6
dm-haiku==0.0.12
dm-tree==0.1.8
dm_control==1.0.20
e3nn-jax==0.20.6
einops==0.8.0
etils==1.7.0
Farama-Notifications==0.0.4
flax==0.8.4
fonttools==4.53.0
fsspec==2024.6.0
glfw==2.7.0
grpcio==1.64.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.11.0
idna==3.7
imageio==2.34.1
importlib-metadata==4.13.0
importlib_resources==6.4.0
jax==0.4.23
jax-md==0.2.8
jaxlib==0.4.23
jaxopt==0.8.3
jmp==0.0.4
jraph==0.0.6.dev0
kiwisolver==1.4.5
labmaze==1.0.6
lxml==5.2.2
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdtraj==1.10.0
mdurl==0.1.2
ml-dtypes==0.4.0
ml_collections==0.1.1
mpmath==1.3.0
msgpack==1.0.8
mujoco==3.1.6
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.15
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pygame==2.5.2
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rdkit==2023.9.6
requests==2.32.3
rich==13.7.1
scipy==1.12.0
Shimmy==0.2.1
six==1.16.0
swig==4.2.1
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorstore==0.1.61
toolz==0.12.1
tqdm==4.66.4
tree-math==0.2.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.2
